pyproject.toml
[project]
name = "personal_finances"
version = "0.1.0"

[tool.poetry]
name = "personal_finances"
readme = "README.md"

package-mode = false

[tool.poetry.dependencies]
python = ">=3.12,<3.14"
requests = "^2.32.3"
python-dotenv = "^1.1.0"
boto3 = "^1.41.5"
boto3-stubs = {extras = ["s3", "ssm"], version = "^1.41.5"}
inflection = "^0.5.1"
alembic = "^1.17.2"
pandas = "^2.3.3"
pyarrow = "^22.0.0"
duckdb = "^1.4.2"
dbt-core = "^1.10.15"
dbt-duckdb = "^1.10.0"
ruff = "^0.14.9"
sqlalchemy = "^2.0.41"
psycopg2-binary = "^2.9.9"
streamlit = "^1.45.1"
streamlit-aggrid = "^1.1.5.post1"
plotly = "^6.1.2"
dash = "^3.0.4"
dash-bootstrap-components = "^2.0.3"
dagster = "^1.12.4"
dagster-webserver = "^1.12.4"
dagster-graphql = "^1.12.4"
dagster-aws = "^0.28.4"
dagster-postgres = "^0.28.4"
dagster-dbt = "^0.28.4"
fastapi = "^0.128.0"
uvicorn = {extras = ["standard"], version = "^0.40.0"}
pytest = "^9.0.2"

[tool.poetry.group.dev]
optional = true

[tool.poetry.group.dev.dependencies]
# Development dependencies
ruff = "^0.14.9"
pre-commit = "^4.5.0"
mypy = "^1.13.0"
sqlfluff = "^3.5.0"
sqlfluff-templater-dbt = "^3.5.0"
coverage = "^7.13.0"
types-psycopg2 = "^2.9.21.20251012"
boto3-stubs = {extras = ["bedrock-runtime", "s3", "ssm"], version = "^1.42.16"}
types-requests = "^2.32.4.20250913"
types-python-dateutil = "^2.9.0.20251115"
types-croniter = "^6.0.0.20250809"
pytest = "^9.0.2"
pytest-cov = "^7.0.0"
pytest-mock = "^3.15.1"


[tool.ruff]
# Ruff Configuration
# Ruff is an extremely fast Python linter and formatter written in Rust
line-length = 100                # Maximum line length for code formatting
target-version = "py313"         # Target Python version for syntax compatibility
exclude = [
    "alembic",
]

[tool.ruff.format]
quote-style = "double"           # Use double quotes for strings
indent-style = "space"           # Use spaces for indentation

# Ruff Linting Rules
# Selects which linting rules to enable and ignore
[tool.ruff.lint]
select = [
    "E",     # pycodestyle errors
    "F",     # pyflakes
    "N",     # pep8-naming
    "D",     # pydocstyle (docstring conventions)
    "PD",    # pandas-vet
    "RUF",   # Ruff-specific rules
    "PL",    # Pylint
    "NPY",   # NumPy-specific rules
    "PTH",   # flake8-use-pathlib
    "RET",   # flake8-return
    "SIM",   # flake8-simplify
    "I",     # isort (import sorting)
    "UP"     # pyupgrade (upgrade syntax for newer Python)
]
ignore = [
    "D213",  # Multi-line docstring summary should start at the second line
    "D203",  # One blank line required before class docstring
    "E501"   # Line too long (handled by formatter)
]

per-file-ignores = { "__init__.py" = ["D"], "testing/**/*.py" = ["PLR2004"] }  # PLR2004: Magic values OK in tests

dummy-variable-rgx = "^(_+|(_+[a-zA-Z0-9_]*[a-zA-Z0-9]+?))$"  # Regex for unused variables

[tool.ruff.lint.mccabe]
max-complexity = 10              # Maximum cyclomatic complexity threshold

# MyPy Configuration
# Static type checking configuration for enhanced type safety
[tool.mypy]
exclude = "(^testing/|^alembic/)"  # Don't type check testing or alembic directories
warn_redundant_casts = true      # Warn about unnecessary casts
warn_unused_ignores = true       # Warn about unused type ignore comments

disallow_incomplete_defs = true  # Functions must have type annotations
disallow_untyped_defs = true     # All functions must have type annotations
disallow_any_generics = true     # Require explicit types for generics

check_untyped_defs = true        # Type check bodies of untyped functions
no_implicit_reexport = true      # Require explicit re-exports

ignore_missing_imports = true    # Don't error on missing type stubs for third-party libraries

# Pytest Configuration
[tool.pytest.ini_options]
testpaths = ["testing"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --tb=short"

# Coverage Configuration
[tool.coverage.run]
source = ["src"]
omit = [
    "*/testing/*",
    "*/__pycache__/*",
    # Infrastructure files requiring running services - tested via integration
    "src/orchestration/*",      # Dagster jobs, assets, schedules
    "src/streamlit/*",          # Streamlit UI pages
    "src/dash_dashboards/*",    # Dash dashboards
    "src/postgres/gocardless/seed.py",  # One-off seed script
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
]

[tool.sqlfluff.core]
templater = "dbt"
dialect = "duckdb"

# Optional but sensible defaults
sql_file_exts = ".sql,.sql.j2,.dml,.ddl"
max_line_length = 120
ignore_templated_areas = true

[tool.sqlfluff.templater.dbt]
project_dir = "dbt"
profiles_dir = "dbt"
profile = "duckdb_local"
